moved files around and gitignored csv folder
