+++ .gitignore
@@ -10,6 +10,8 @@ pids
 *.pid
 logs
 *.log
+
+/csv
 *.csv
 # Directory for instrumented libs generated by jscoverage/JSCover
 lib-cov
