+++ csvGenerator.js
@@ -4,47 +4,74 @@ const faker = require('faker');
 const debug = require('debug')('app:gen:psql');
 require('events').EventEmitter.defaultMaxListeners = 100000;
 
-const usersWriter = csvWriter();
-const listingsWriter = csvWriter();
-const bookingsWriter = csvWriter();
-
 const getRandomNum = (min, max) => Math.floor((Math.random() * (max - min) ) + min);
 
-const generateUsers = async () => {
-  usersWriter.pipe(fs.createWriteStream('./csv/users.csv'));
-  for(let i = 1; i <= 1000; i++) {
-    await usersWriter.write({
-      id: i,
-      name: faker.name.findName(),
-    });
-  }
-  usersWriter.end();
-  console.log('Users generated!');
-};
+const writeUsers = fs.createWriteStream('./csv/users.csv');
+writeUsers.write('id,name\n', 'utf8');
 
-const generateListings = async () => {
-  listingsWriter.pipe(fs.createWriteStream('./csv/listings.csv'));
-  let reservationId = 0;
-  const startDate = new Date();
-  for (let i = 1; i <= 1000; i++) {
-    const listingInfo = {
-      id: i,
-      name: faker.name.findName(),
-      max_guests: getRandomNum(1, 17),
-      max_stay: getRandomNum(2, 32),
-      review_count: getRandomNum(10, 501),
-      per_night: getRandomNum(50, 801),
-      cleaning: i % 3 === 0 ? getRandomNum(10, 101) : 0,
-      service: getRandomNum(20, 101),
+function writeTenMillionUsers(writer, encoding, callback) {
+  let i = 100;
+  let id = 0;
+  write();
+  function write() {
+    let ok = true;
+    do {
+      i--;
+      id++;
+      const name = faker.name.findName();
+      const data = `${id},${name}\n`;
+      if (i === 0) {
+        writer.write(data, encoding, callback);
+      } else {
+        ok = writer.write(data, encoding);
+      }
+    } while (i > 0 && ok);
+    if (i > 0) {
+      writer.once('drain', write);
     }
-    await listingsWriter.write(listingInfo);
-    generateBookings(reservationId + 1, listingInfo.review_count, startDate, listingInfo);
-    reservationId += listingInfo.review_count;
   }
-  listingsWriter.end();
-  console.log('Listings and bookings generated!');
-};
+}
 
+const writeListings = fs.createWriteStream('./csv/listings.csv');
+writeListings.write('id,name,max_guests,max_stay,review_count,per_night,cleaning,service\n', 'utf8');
+
+function writeOneMillionListings(writer, encoding, callback) {
+  let i = 10;
+  let id = 0;
+  write();
+  function write() {
+    let ok = true;
+    let reservationId = 0;
+    const startDate = new Date();
+    do {
+      i--;
+      id++;
+      const listingInfo = {
+        id,
+        name: faker.name.findName(),
+        max_guests: getRandomNum(1, 17),
+        max_stay: getRandomNum(2, 32),
+        review_count: getRandomNum(10, 100),
+        per_night: getRandomNum(50, 801),
+        cleaning: i % 3 === 0 ? getRandomNum(10, 101) : 0,
+        service: getRandomNum(20, 101),
+      };
+      const data = `${listingInfo.id},${listingInfo.name},${listingInfo.max_guests},${listingInfo.max_stay},${listingInfo.review_count},${listingInfo.per_night},${listingInfo.cleaning},${listingInfo.service}\n`;
+      generateBookings(reservationId + 1, listingInfo.review_count, startDate, listingInfo);
+      reservationId += listingInfo.review_count;
+      if (i === 0) {
+        writer.write(data, encoding, callback);
+      } else {
+        ok = writer.write(data, encoding);
+      }
+    } while (i > 0 && ok);
+    if (i > 0) {
+      writer.once('drain', write);
+    }
+  }
+}
+
+const bookingsWriter = csvWriter();
  const generateBookings = async (startingId, reviewCount, startDate, listingInfo) => {
    bookingsWriter.pipe(fs.createWriteStream('./csv/bookings.csv'));
 
@@ -75,20 +102,16 @@ const generateListings = async () => {
        total_cost: (listingInfo.per_night * stayLength) + listingInfo.cleaning + listingInfo.service,
        listing_id: listingInfo.id,
        user_id: getRandomNum(1, 1000),
-     }, () => { 
-       bookingsWriter.end();
-       bookingsWriter.on('finish', () => {
-         debug('done');
-       })
-     });
+     }, () => { bookingsWriter.end(); });
      checkin = calculateDate(checkin, stayLength);
    }
  };
 
-async function dataGenerator() {
+const generateData = async () => {
   debug('start');
-  await generateUsers();
-  await generateListings();
-}
+  await writeTenMillionUsers(writeUsers, 'utf8', () => { writeUsers.end(); });
+  await writeOneMillionListings(writeListings, 'utf8', () => { writeUsers.end(); });
+  debug('done');
+};
 
-dataGenerator();
+generateData();
